refactor: structure according to chain, fix tests
